rust: remove RustContext required to link up functions
Instead Rust will attempt to link up some function pointers on first
call. This is done by wrapping the resolution in lazy_static, where
on first use, dlsym() will be used to lookup the address of the
function on first use.

This eliminates the requirement of passing a RustContext into the
rust code, but more importantly this allows plugins to be written
in Rust without requiring the plugin to execute the same boilerplate
context loading as the main process.

Ideally none of this would be required, but remains hard until we
have a standalone libsuricata, that for example, the Rust unit tests
could use.  But some glue like this will probably always be required
as a Rust plugin use the log crate, has its log crate in a different
address space than the log crate of the main process, so some
dynamic link is still required.